requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
et-xmlfile==1.0.1
fuzzywuzzy==0.18.0
google-auth==1.21.2
google-auth-oauthlib==0.4.1
gspread==3.6.0
idna==2.10
idna-ssl==1.1.0
jdcal==1.4.1
multidict==4.7.6
oauthlib==3.1.0
openpyxl==3.0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.5.1
